feat(TextExtractorKBert): extract to csv
