+++ matching-jena-matchers/src/main/java/de/uni_mannheim/informatik/dws/melt/matching_jena_matchers/util/textExtractors/kBert/processedNode/ProcessedRDFNode.java
@@ -1,6 +1,8 @@
-package de.uni_mannheim.informatik.dws.melt.matching_jena_matchers.util.textExtractorsMap.kBert.processedNode;
+package de.uni_mannheim.informatik.dws.melt.matching_jena_matchers.util.textExtractors.kBert.processedNode;
 
 import de.uni_mannheim.informatik.dws.melt.matching_jena_matchers.util.StringProcessing;
+
+import static org.apache.commons.text.StringEscapeUtils.escapeCsv;
 
 public abstract class ProcessedRDFNode {
 
@@ -10,7 +12,7 @@ public abstract class ProcessedRDFNode {
      * @return todo
      */
     public String getNormalized() {
-        return StringProcessing.normalizeJoining(getRaw());
+        return escapeCsv(StringProcessing.normalizeJoining(getRaw()));
     }
 
     protected abstract String getRaw();
