+++ matching-jena-matchers/src/test/java/de/uni_mannheim/informatik/dws/melt/matching_jena_matchers/util/textExtractors/KBertSentenceTransformersMatcherTest.java
@@ -1,9 +1,9 @@
-package de.uni_mannheim.informatik.dws.melt.matching_jena_matchers.util.textExtractorMap;
+package de.uni_mannheim.informatik.dws.melt.matching_jena_matchers.util.textExtractors;
 
 import de.uni_mannheim.informatik.dws.melt.matching_base.FileUtil;
 import de.uni_mannheim.informatik.dws.melt.matching_data.TestCase;
 import de.uni_mannheim.informatik.dws.melt.matching_data.TrackRepository;
-import de.uni_mannheim.informatik.dws.melt.matching_jena_matchers.util.textExtractorsMap.kBert.TextExtractorKBert;
+import de.uni_mannheim.informatik.dws.melt.matching_jena_matchers.util.textExtractors.kBert.TextExtractorKBert;
 import de.uni_mannheim.informatik.dws.melt.matching_ml.python.nlptransformers.kbert.KBertSentenceTransformersMatcher;
 import org.apache.jena.ontology.OntModel;
 import org.junit.jupiter.api.Test;
@@ -26,7 +26,7 @@ public class KBertSentenceTransformersMatcherTest {
         OntModel sourceOntology = getTransformedObject(source, OntModel.class, properties);
         KBertSentenceTransformersMatcher matcher = new KBertSentenceTransformersMatcher(
                 new TextExtractorKBert(), "paraphrase-MiniLM-L6-v2");
-        File corpus = FileUtil.createFileWithRandomNumber("corpus", ".txt");
+        File corpus = FileUtil.createFileWithRandomNumber("corpus", ".csv");
 
         // when
         matcher.createTextFile(sourceOntology, corpus, matcher.getResourcesExtractor().get(0), properties);
